unify unit test utilities in single file
